requirements.txt
langchain
langchain-community
openai
matplotlib
plotly
streamlit
python-dotenv